--- src/components/Pacs/components/StudyCardCopy.tsx
@@ -1,267 +1,0 @@
-import {
-  Badge,
-  Button,
-  Card,
-  CardHeader,
-  Grid,
-  GridItem,
-  Skeleton,
-  Tooltip,
-} from "@patternfly/react-core";
-import { Alert } from "antd";
-import { useContext, useEffect, useState } from "react";
-import { DotsIndicator } from "../../Common";
-import {
-  DownloadIcon,
-  PreviewIcon,
-  QuestionCircleIcon,
-  ThLargeIcon,
-} from "../../Icons";
-import { PacsQueryContext, Types } from "../context";
-import SeriesCard from "./SeriesCardCopy";
-import { CardHeaderComponent } from "./SettingsComponents";
-import { formatStudyDate } from "./utils";
-//import useInterval from "./useInterval";
-import useSettings from "../useSettings";
-
-const StudyCardCopy = ({ study }: { study: any }) => {
-  const { data, isLoading, error } = useSettings();
-  const { state, dispatch } = useContext(PacsQueryContext);
-  const [isStudyExpanded, setIsStudyExpanded] = useState(false);
-  const { seriesPreviews, preview, pullStudy, studyPullTracker } = state;
-  const userPreferences = data?.study;
-  const userPreferencesArray = userPreferences && Object.keys(userPreferences);
-  const studyInstanceUID = study.StudyInstanceUID.value;
-
-  useEffect(() => {
-    if (studyPullTracker && pullStudy) {
-      const studyBeingTracked = studyPullTracker[studyInstanceUID];
-      if (studyBeingTracked) {
-        let allSeriesBeingTracked = true;
-        for (const series in studyBeingTracked) {
-          const isSeriesDone = studyBeingTracked[series];
-          if (!isSeriesDone) {
-            allSeriesBeingTracked = false;
-            break;
-          }
-        }
-        if (allSeriesBeingTracked) {
-          dispatch({
-            type: Types.SET_PULL_STUDY,
-            payload: null,
-          });
-        }
-      }
-    }
-  }, [studyPullTracker, pullStudy]);
-
-  return (
-    <>
-      <Card isExpanded={isStudyExpanded} isRounded isSelectable isClickable>
-        <CardHeader
-          actions={{
-            actions: <CardHeaderComponent resource={study} type="study" />,
-          }}
-          className="flex-studies-container"
-          onExpand={() => setIsStudyExpanded(!isStudyExpanded)}
-        >
-          {isLoading ? (
-            <GridItem lg={4} md={4} sm={12}>
-              <Skeleton
-                width="100%"
-                height="100%"
-                screenreaderText="Loading contents"
-              />
-            </GridItem>
-          ) : error ? (
-            <GridItem>
-              <Alert type="error" description="Please Refresh the page..." />
-            </GridItem>
-          ) : (
-            <>
-              {userPreferences &&
-              userPreferencesArray &&
-              userPreferencesArray.length > 0 ? (
-                userPreferencesArray.map((key: string) => (
-                  <div key={key} className="flex-studies-item">
-                    <div className="study-detail-title">{key}</div>
-                    <Tooltip content={study[key].value} position="auto">
-                      <div
-                        style={{
-                          overflow: "hidden",
-                          textOverflow: "ellipsis",
-                          whiteSpace: "nowrap",
-                        }}
-                      >
-                        <span style={{ marginRight: "0.5em" }}>
-                          {study[key].value && study[key].value}
-                        </span>{" "}
-                      </div>
-                    </Tooltip>
-                  </div>
-                ))
-              ) : (
-                <>
-                  <div className="flex-studies-item">
-                    <Tooltip
-                      content={study.StudyDescription.value}
-                      position="auto"
-                    >
-                      <div
-                        style={{
-                          overflow: "hidden",
-                          textOverflow: "ellipsis",
-                          whiteSpace: "nowrap",
-                        }}
-                      >
-                        <span style={{ marginRight: "0.5em" }}>
-                          {study.StudyDescription.value &&
-                            study.StudyDescription.value}
-                        </span>{" "}
-                      </div>
-                    </Tooltip>
-                    <div>
-                      {study.NumberOfStudyRelatedSeries.value &&
-                        study.NumberOfStudyRelatedSeries.value}{" "}
-                      series, {`${formatStudyDate(study.StudyDate.value)}`}
-                    </div>
-                  </div>
-                  <div className="flex-studies-item ">
-                    <div className="study-detail-title">
-                      Modalities in Study
-                    </div>
-                    <div>
-                      {study.ModalitiesInStudy.value
-                        ?.split("\\")
-                        .map((m: string, index: number) => (
-                          <Badge
-                            style={{ margin: "auto 0.125em" }}
-                            key={`${m} _${index}`}
-                          >
-                            {m}
-                          </Badge>
-                        ))}
-                    </div>
-                  </div>
-                  <div className="flex-studies-item">
-                    <div className="study-detail-title">Accession Number</div>
-                    {study.AccessionNumber.value?.startsWith("no value") ? (
-                      <Tooltip content={study.AccessionNumber}>
-                        <QuestionCircleIcon />
-                      </Tooltip>
-                    ) : (
-                      <div>{study.AccessionNumber.value}</div>
-                    )}
-                  </div>
-                  <div className="flex-studies-item">
-                    <div className="study-detail-title">Station</div>
-                    {study.PerformedStationAETitle.value?.startsWith(
-                      "no value",
-                    ) ? (
-                      <Tooltip content={study.PerformedStationAETitle.value}>
-                        <QuestionCircleIcon />
-                      </Tooltip>
-                    ) : (
-                      <div>{study.PerformedStationAETitle.value}</div>
-                    )}
-                  </div>
-                </>
-              )}
-              <div className="flex-studies-item button-container">
-                {import.meta.env.VITE_OHIF_URL && (
-                  <Tooltip content="Open in OHIF">
-                    <Button
-                      size="sm"
-                      variant="secondary"
-                      style={{ marginRight: "0.25em" }}
-                      icon={<ThLargeIcon />}
-                      component="a"
-                      href={`${
-                        import.meta.env.VITE_OHIF_URL
-                      }viewer?StudyInstanceUIDs=${
-                        study.StudyInstanceUID.value
-                      }`}
-                      target="_blank"
-                    />
-                  </Tooltip>
-                )}
-                <Tooltip
-                  content={
-                    preview === true ? "Hide All Previews" : "Show All Previews"
-                  }
-                >
-                  <Button
-                    size="sm"
-                    variant="tertiary"
-                    style={{ marginRight: "0.25em" }}
-                    onClick={() => {
-                      dispatch({
-                        type: Types.SET_SHOW_PREVIEW,
-                        payload: {
-                          preview: !preview,
-                        },
-                      });
-
-                      if (preview === false) {
-                        setIsStudyExpanded(true);
-                      }
-                    }}
-                    icon={<PreviewIcon />}
-                  />
-                </Tooltip>
-
-                {pullStudy ? (
-                  <DotsIndicator title="Pulling Study..." />
-                ) : (
-                  <Tooltip content="Pull Study">
-                    <Button
-                      onClick={() => {
-                        dispatch({
-                          type: Types.SET_PULL_STUDY,
-                          payload: null,
-                        });
-                        setIsStudyExpanded(true);
-                      }}
-                      variant="tertiary"
-                      className="button-with-margin"
-                      size="sm"
-                      icon={<DownloadIcon />}
-                    />
-                  </Tooltip>
-                )}
-              </div>
-            </>
-          )}
-        </CardHeader>
-      </Card>
-      {isStudyExpanded && (
-        <div className="patient-series">
-          <Grid hasGutter>
-            {study.series.map((series: any) => {
-              const seriesID = series.SeriesInstanceUID.value;
-              const seriesPreview = seriesPreviews?.[seriesID];
-              return (
-                <GridItem
-                  key={series.SeriesInstanceUID.value}
-                  className={seriesPreviews?.[seriesID] ? "series-grid" : ""}
-                  rowSpan={1}
-                  lg={seriesPreview ? 4 : 12}
-                  md={seriesPreview ? 4 : 12}
-                  sm={12}
-                  xl={seriesPreview ? 3 : 12}
-                >
-                  <SeriesCard
-                    key={series.SeriesInstanceUID.value}
-                    series={series}
-                  />
-                </GridItem>
-              );
-            })}
-          </Grid>
-        </div>
-      )}
-    </>
-  );
-};
-
-export default StudyCardCopy;
