--- src/components/Pacs/components/useInterval.tsx
@@ -1,27 +1,0 @@
-import { useEffect, useRef } from "react";
-
-export default function useInterval(
-  callback: () => void,
-  delay: number | null,
-): void {
-  const savedCallback = useRef<() => void | null>();
-
-  useEffect(() => {
-    savedCallback.current = callback;
-  }, [callback]);
-
-  useEffect(() => {
-    function tick() {
-      if (savedCallback.current) {
-        savedCallback.current();
-      }
-    }
-
-    if (delay !== null) {
-      const id = setInterval(tick, delay);
-      return () => {
-        clearInterval(id);
-      };
-    }
-  }, [delay]);
-}
